UNIT_4/src/main/java/org/example/BookService.java
package org.example;

// Press Shift twice to open the Search Everywhere dialog and type `show whitespaces`,
// then press Enter. You can now see whitespace characters in your code.
public class BookService {
    private BookRepository bookRepository;

    public BookService(BookRepository bookRepository) {
        this.bookRepository = bookRepository;
    }

    public int getBookCount() {
        return bookRepository.getCount();
    }

    // Другие методы BookService, использующие bookRepository
}

